changing wards uint to uint256

+++ src/dog.sol
@@ -51,7 +51,7 @@ interface VowLike {
 
 contract Dog {
     // --- Auth ---
-    mapping (address => uint) public wards;
+    mapping (address => uint256) public wards;
     function rely(address usr) external auth { wards[usr] = 1; emit Rely(usr); }
     function deny(address usr) external auth { wards[usr] = 0; emit Deny(usr); }
     modifier auth {
